requirements.txt
fastapi
uvicorn[standard]
pyjwt
python-dotenv
python-multipart
elasticsearch
langchain
langchain-google-genai
bcrypt
pytest
pytest-asyncio
